requirements.txt
streamlit>=1.41.0
pandas>=2.2.0
plotly>=5.18.0
google-genai>=1.0.0
python-dotenv>=1.0.0
groq>=0.13.0
